[NEW-50432] fix: maps overwriting dashboards (#2139)
* fixing map refactor

* Advanced Editor > update change function

* revert equal number opt in due to color dif.

* josh - changed hooks to helpers

* fix: scaling and color palette issue

* fix: generateRuntimeData import on world maps

* fix: setRuntimeData calls

---------
