+++ packages/map/src/CdcMapComponent.tsx
@@ -41,6 +41,8 @@ import {
   hashObj,
   navigationHandler
 } from './helpers'
+import generateRuntimeLegend from './helpers/generateRuntimeLegend'
+import generateRuntimeData from './helpers/generateRuntimeData'
 
 // Child Components
 import Annotation from './components/Annotation'
@@ -56,8 +58,6 @@ import GoogleMap from './components/GoogleMap'
 
 // hooks
 import useResizeObserver from './hooks/useResizeObserver'
-import useGenerateRuntimeLegend from './hooks/useGenerateRuntimeLegend'
-import useGenerateRuntimeData from './hooks/useGenerateRuntimeData'
 import { VizFilter } from '@cdc/core/types/VizFilter'
 import { getInitialState, mapReducer } from './store/map.reducer'
 import { RuntimeData } from './types/RuntimeData'
@@ -68,8 +68,8 @@ import useModal from './hooks/useModal'
 
 type CdcMapComponent = {
   config: MapConfig
-  isDashboard?: boolean
   isEditor?: boolean
+  isDashboard?: boolean
   link?: string
   logo?: string
   navigationHandler: Function
@@ -116,18 +116,15 @@ const CdcMapComponent: React.FC<CdcMapComponent> = ({
 
   const editorContext = useContext(EditorContext)
 
-  const setConfig = (newMapState: MapConfig): void => {
-    dispatch({ type: 'SET_CONFIG', payload: newMapState })
-
+  const setConfig = (newMapConfig: MapConfig): void => {
+    dispatch({ type: 'SET_CONFIG', payload: newMapConfig })
     if (isEditor && !isDashboard) {
-      editorContext.setTempConfig(newMapState)
-    } else if (isDashboard && isEditor) {
-      setParentConfig(newMapState)
+      editorContext.setTempConfig(newMapConfig)
     }
   }
 
   useEffect(() => {
-    const _newConfig = _.cloneDeep(configObj)
+    const _newConfig = getInitialState(_.cloneDeep(configObj)).config
     if (configObj.data) {
       _newConfig.data = configObj.data
     }
@@ -167,8 +164,6 @@ const CdcMapComponent: React.FC<CdcMapComponent> = ({
 
   // hooks
   const { currentViewport, dimensions, container, outerContainerRef } = useResizeObserver(isEditor)
-  const { generateRuntimeLegend } = useGenerateRuntimeLegend(legendMemo, legendSpecialClassLastMemo)
-  const { generateRuntimeData } = useGenerateRuntimeData(config)
 
   const reloadURLData = async () => {
     if (config.dataUrl) {
@@ -283,15 +278,25 @@ const CdcMapComponent: React.FC<CdcMapComponent> = ({
 
     // Data
     if (hashData !== runtimeData?.fromHash && config.data?.fromColumn) {
+      const isCategoryLegend = config?.legend?.type === 'category'
       const newRuntimeData = generateRuntimeData(
         { ...config, data: configObj.data },
         filters || runtimeFilters,
-        hashData
+        hashData,
+        isCategoryLegend
       )
       setRuntimeData(newRuntimeData)
     } else {
       if (hashLegend !== runtimeLegend?.fromHash && undefined === runtimeData?.init) {
-        const legend = generateRuntimeLegend(config, runtimeData, hashLegend)
+        const legend = generateRuntimeLegend(
+          config,
+          runtimeData,
+          hashLegend,
+          setConfig,
+          runtimeFilters,
+          legendMemo,
+          legendSpecialClassLastMemo
+        )
         setRuntimeLegend(legend)
       }
     }
@@ -299,20 +304,17 @@ const CdcMapComponent: React.FC<CdcMapComponent> = ({
 
   useEffect(() => {
     const hashLegend = generateRuntimeLegendHash(config, runtimeFilters)
-    const legend = generateRuntimeLegend({ ...config, data: configObj.data }, runtimeData, hashLegend)
+    const legend = generateRuntimeLegend(
+      { ...config, data: configObj.data },
+      runtimeData,
+      hashLegend,
+      setConfig,
+      runtimeFilters,
+      legendMemo,
+      legendSpecialClassLastMemo
+    )
     setRuntimeLegend(legend)
-  }, [
-    runtimeData,
-    config.legend.unified,
-    config.legend.showSpecialClassesLast,
-    config.legend.separateZero,
-    config.general.equalNumberOptIn,
-    config.legend.numberOfItems,
-    config.legend.specialClasses,
-    config.legend.additionalCategories,
-    config,
-    runtimeFilters
-  ])
+  }, [runtimeData, config, runtimeFilters])
 
   useEffect(() => {
     if (!isDashboard) {
